refactor has-break to keep external hasbreak code unchanged
