+++ lib/logic/tools/crypto_and_encodings/wherigo_urwigo/krevo.wherigotools/ucommons.dart
@@ -24,7 +24,7 @@ int RSHash(String string) {
   return hash;
 }
 
-String _convBase(int numberInput, String fromBaseInput, String toBaseInput) {
+String convBase(int numberInput, String fromBaseInput, String toBaseInput) {
   var toBase = toBaseInput.split('').toList();
   var toLen = toBaseInput.length;
   var retval = '';
@@ -43,10 +43,10 @@ String _convBase(int numberInput, String fromBaseInput, String toBaseInput) {
 /*
   This function will yield collisions for the desired hash
 */
-String findHashAlphaNumeric(int hashToFind, int len) {
+String findHash(int hashToFind, int len) {
   var max = pow(26, len);
   for (var i = 0; i < max; i++) {
-    var s = _convBase(i, '0123456789', 'abcdefghijklmnopqrstuvwxyz').padLeft(len, 'a');
+    var s = convBase(i, '0123456789', 'abcdefghijklmnopqrstuvwxyz').padLeft(len, 'a');
 
     if (RSHash(s) == hashToFind) {
       return s;
@@ -56,15 +56,3 @@ String findHashAlphaNumeric(int hashToFind, int len) {
   return null;
 }
 
-String findHashNumeric(int hashToFind, int len) {
-  var max = pow(26, len);
-  for (var i = 0; i < max; i++) {
-    var s = _convBase(i, '0123456789', '01234567890').padLeft(len, '0');
-
-    if (RSHash(s) == hashToFind) {
-      return s;
-    }
-  }
-
-  return null;
-}
